sprint-4/step-1 feat: типизировать хук UseTabTypes
